requirements.txt
Flask==2.0.3
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.4
Pillow==8.4.0
python-pptx==0.6.21
PyMuPDF==1.19.6
